1. extract ProfileList from ProfilePage; 2. combine ProfileItem into ProfileList;
