--- src/pages/OfferModal/ProfilePage/ProfileItem/index.js
@@ -1,1 +1,0 @@
-export { default } from './ProfileItem';
