+++ src/pages/OfferModal/ProfilePage/ProfileList/ProfileList.jsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 
-import styles from './ProfileItem.module.scss';
+import styles from './ProfileList.module.scss';
+
+import createBirthdayLabel from './utils/createBirthdayLabel';
+import FORM from '../form';
 
 const cx = classNames.bind(styles);
 
 const checkIcon = String.fromCharCode(10003);
 const arrowIcon = String.fromCharCode(10095);
 
-export default function ProfileItem({
+function ProfileItem({
   itemName, handleClick, filled, statusLabel,
 }) {
 
@@ -38,6 +41,34 @@ export default function ProfileItem({
           {arrowIcon}
         </div>
       </button>
+    </div>
+  );
+}
+
+export default function ProfileList({
+  formData, clickHandlers,
+}) {
+  return (
+    <div className={styles.profile_list_wrapper} >
+      {Object.keys(FORM).map((key) => {
+        const { label } = FORM[key];
+        const value = formData[key];
+
+        const statusLabel = {
+          birthday: createBirthdayLabel(value),
+          mobile: value,
+        }[key] || null;
+
+        return (
+          <ProfileItem
+            itemName={label}
+            handleClick={clickHandlers[key]}
+            statusLabel={statusLabel}
+            filled={!!value}
+            key={key}
+          />
+        );
+      })}
     </div>
   );
 }
