+++ src/pages/OfferModal/ProfilePage/ProfileList/ProfileList.module.scss
@@ -54,3 +54,10 @@
     }
   }
 }
+
+.profile_list_wrapper{
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+}
+
