+++ src/pages/OfferModal/ProfilePage/ProfileList/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProfileList';
