+++ src/pages/OfferModal/ProfilePage/ProfilePage.module.scss
@@ -15,12 +15,6 @@
 .content_wrapper{
   margin: 0 15px;
   width: 360px;
-  
-  .profile_list_wrapper{
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-  }
 }
 
 .back_button_wrapper{
